requirements.txt
streamlit==1.28.1
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
requests==2.31.0
huggingface-hub==0.19.4